requirements.txt
Pillow == 8.1.0
PyYAML == 5.3.1
SoundFile == 0.10.3.post1
espnet_model_zoo == 0.0.0a25
ipython == 7.20.0
librosa == 0.8.0
matplotlib == 3.1.0
numpy == 1.20.0
pandas == 1.2.2
pytorch_lightning == 1.1.7
requests == 2.25.1
scikit_image == 0.18.1
scikit_learn == 0.24.1
torch == 1.7.1
torchvision == 0.8.2
tqdm == 4.56.0
umap_learn == 0.5.1
more-itertools == 8.7.0 